+++ src/main/java/dev/yatloaf/modkrowd/config/Predicate.java
@@ -3,6 +3,7 @@ package dev.yatloaf.modkrowd.config;
 import dev.yatloaf.modkrowd.cubekrowd.common.CKColor;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.Subserver;
 import dev.yatloaf.modkrowd.cubekrowd.subserver.TeamSet;
+import net.minecraft.client.gui.components.Tooltip;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.Style;
 
@@ -22,13 +23,13 @@ public class Predicate {
 
     public final String id;
     public final Component name;
-    public final Component tooltip;
+    public final Tooltip tooltip;
     public final Test test;
 
     public Predicate(String id, Style style, Test test) {
         this.id = id;
         this.name = Component.translatable("modkrowd.config.predicate." + id).setStyle(style);
-        this.tooltip = Component.translatable("modkrowd.config.predicate." + id + ".tooltip");
+        this.tooltip = Tooltip.create(Component.translatable("modkrowd.config.predicate." + id + ".tooltip"));
         this.test = test;
         FROM_ID.put(id, this);
     }